+++ libs/fuzzy-checker/src/vivarium/fuzzy_checker/__init__.py
@@ -1,6 +1,13 @@
-#################
-# Fuzzy Checker #
-#################
+"""
+======================
+Vivarium Fuzzy Checker
+======================
+
+The ``FuzzyChecker`` is a tool for statistical "fuzzy" checks of values subject
+to stochastic variation and used to verify and validate Vivarium simulation
+outputs.
+"""
+
 from importlib.metadata import PackageNotFoundError, version
 
 try: