+++ libs/artifact/CHANGELOG.rst
@@ -1,3 +1,8 @@
+**1.0.8 - 07/23/26**
+
+- Streamline package __init__.py docstring
+- Add project.urls to pyproject.toml
+
 **1.0.7 - 07/14/26**
 
 - Update stale references for monorepo libraries