+++ libs/risk-distributions/CHANGELOG.rst
@@ -1,3 +1,8 @@
+**3.1.8 - 07/23/26**
+
+- Streamline package __init__.py docstring
+- Add project.urls to pyproject.toml
+
 **3.1.7 - 07/14/26**
 
 - Update stale references for monorepo libraries